# perf: plugin update and load time