--- a/ape_ens/__init__.py
+++ b/ape_ens/__init__.py
@@ -1,9 +1,19 @@
 from ape import plugins
-from ape.types import AddressType
-
-from .converter import ENSConversions
 
 
 @plugins.register(plugins.ConversionPlugin)
 def converters():
+    from ape.types import AddressType
+
+    from ape_ens.converter import ENSConversions
+
     yield AddressType, ENSConversions
+
+
+def __getattr__(name: str):
+    if name == "ENSConversions":
+        from ape_ens.converter import ENSConversions
+
+        return ENSConversions
+
+    raise AttributeError(name)